adding

+++ ch03/main.c
@@ -119,11 +119,11 @@ int main(void)
 
       			 //GPIOA->BSRR = GPIO_PIN_5; //ON
       			 //  GPIOA->BSRR = ((uint16_t)0x0020); //ON
-      			 GPIOA->BSRR = ((uint16_t)0b0000000000100000); // ON
+      			 GPIOA->BSRR = ((uint32_t)0b0000000000100000); // ON
 
       		 }else if(!flag_set1){
       		    // GPIOA->BSRR = ((uint16_t)0x0020) << 16;  //OFF
-      		     GPIOA->BSRR = ((uint16_t)0b0000000000100000) << 16;
+      		     GPIOA->BSRR = ((uint32_t)0b0000000000100000) << 16;
       		 }
 
       	  }else if(!(GPIOC->IDR & GPIO_PIN_13)&& flag_block1){
